REST API course listing/leaving, specific courses, and questions
Added new routes to leave a course and also leave its administration. 
- delete '/leave/course': student can leave a course providing the key of the course.
- delete '/leave_admin': student can leave administration of a course by providing its key.

Updated approval of new administrators by professors. New route takes as parameter the key of the course and the body of the request only takes the student's email and verdict of the professor (boolean).

Added three new routes to get list of courses to which a student is enrolled as well as a list to get basic info a a course given its key. Also, added the route to post new questions to a course given its key.
- get '/courses': given the jwt of a user get a listing of its courses with basic info.
- get '/courses/:course' given the jwt of a user and the course key obtain basic info of the course as well as its questions.
- post '/courses/:course/new_question': given the jwt of a user, the key of a course and a body containing title, content and tags a user can post a question to the specified course. The route creates an identifier for the new question using nanoid, checks if the user is enrolled to the course and if the tags are valid. Next update will bring anonymous posting support by adding a new boolean element to the body.