Routes for Express app
Created 3 new files in /HelpMate/src/backend/routes. The files are: courses.js, login.js and register.js. These are the routes for the express app. Now the routes can be accessed without the '/api' route, to work in conjunction with the react app. Methods still work the same as in the users api.