API get course user lsit
New route '/courses/:course/users' that returns an array with two positions. The first one returns an array of the users that are administrators of the course and the second one an array of the users of the course. The method only expects the jwt of the user and the key of the course.